fix getQuaternion() regs are signed

+++ src/BNO055ESP32.cpp
@@ -464,10 +464,11 @@ bno055_quaternion_t BNO055::getQuaternion(){
 	readLen(BNO055_REG_QUA_DATA_W_LSB, 8, buffer);
 	
 	bno055_quaternion_t wxyz;
-	wxyz.w = ((((uint16_t)buffer[1]) << 8) | ((uint16_t)buffer[0]))/scale;
-	wxyz.x = ((((uint16_t)buffer[3]) << 8) | ((uint16_t)buffer[2]))/scale;
-	wxyz.y = ((((uint16_t)buffer[5]) << 8) | ((uint16_t)buffer[4]))/scale;
-	wxyz.z = ((((uint16_t)buffer[7]) << 8) | ((uint16_t)buffer[6]))/scale;
+	wxyz.w = (int16_t)(((buffer[1]) << 8) | buffer[0])/scale;
+	wxyz.x = (int16_t)(((buffer[3]) << 8) | buffer[2])/scale;
+	wxyz.y = (int16_t)(((buffer[5]) << 8) | buffer[4])/scale;
+	wxyz.z = (int16_t)(((buffer[7]) << 8) | buffer[6])/scale;
+
 	return wxyz;
 }
